requirements.txt
# XFeed Memo — Dependencies
# Install: pip install -r requirements.txt

# Core scraper (existing)
playwright>=1.50

# New pipeline (src/)
PyYAML>=6.0
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
